server/requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.3
Flask_Cors==4.0.0
Flask_PyMongo==2.3.0
nltk==3.8.1
numpy==2.2.1
pymongo==4.10.1
python-dotenv==1.0.0
requests==2.31.0
Sastrawi==1.0.1
scikit-learn==1.3.0
scipy==1.11.3
selenium==4.27.1
